requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=1.6.0=pypi_0
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.5.0=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.8.0=pypi_0
args=0.1.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.1.0=pypi_0
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
chumpy=0.70=pypi_0
clarabel=0.10.0=pypi_0
click=8.1.8=pypi_0
clint=0.5.1=pypi_0
clip=0.2.0=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.1=pypi_0
crcmod=1.7=pypi_0
cryptography=44.0.2=pypi_0
cvxpy=1.6.3=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.12=pypi_0
datasets=3.3.2=pypi_0
diffusers=0.32.2=pypi_0
dill=0.3.8=pypi_0
einops=0.8.1=pypi_0
embreex=2.17.7.post6=pypi_0
exceptiongroup=1.2.2=pypi_0
fairscale=0.4.13=pypi_0
fastapi=0.115.11=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.14.0=pypi_0
fire=0.7.0=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.56.0=pypi_0
freetype-py=2.5.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
ftfy=6.3.1=pypi_0
future=1.0.0=pypi_0
gradio=5.17.1=pypi_0
gradio-client=1.7.1=pypi_0
grpcio=1.71.0=pypi_0
h11=0.14.0=pypi_0
hpsv2=1.2.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
image-reward=1.5=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pypi_0
iniconfig=2.0.0=pypi_0
invisible-watermark=0.2.0=pypi_0
jax=0.5.2=pypi_0
jaxlib=0.5.1=pypi_0
jinja2=3.1.6=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.4.2=pypi_0
json-tricks=3.17.3=pypi_0
jsonlines=4.0.0=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.14.3=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.1=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.10.21=pypi_0
ml-dtypes=0.5.1=pypi_0
mmcv=1.5.0=pypi_0
mmdet=3.3.0=pypi_0
mmengine=0.10.7=pypi_0
model-index=0.1.11=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
munkres=1.1.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
numpy=1.23.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-ml-py=12.570.86=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
nvitop=1.4.2=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.21.0=pypi_0
openai-clip=1.0.1=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
opendatalab=0.0.10=pypi_0
openmim=0.3.9=pypi_0
openssl=3.0.16=h5eee18b_0
openxlab=0.1.2=pypi_0
opt-einsum=3.4.0=pypi_0
ordered-set=4.1.0=pypi_0
orjson=3.10.15=pypi_0
osqp=0.6.7.post3=pypi_0
oss2=2.17.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
peft=0.14.0=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py310h06a4308_0
platformdirs=4.3.7=pypi_0
pluggy=1.5.0=pypi_0
propcache=0.3.0=pypi_0
protobuf=4.25.6=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=19.0.1=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.22.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydeprecate=0.3.1=pypi_0
pydub=0.25.1=pypi_0
pyglet=2.1.3=pypi_0
pygments=2.19.1=pypi_0
pyopengl=3.1.0=pypi_0
pyparsing=3.2.1=pypi_0
pyrender=0.1.45=pypi_0
pytest=7.2.0=pypi_0
pytest-split=0.8.0=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytorch-fid=0.3.0=pypi_0
pytorch-lightning=2.3.1=pypi_0
pytz=2023.4=pypi_0
pywavelets=1.8.0=pypi_0
pyyaml=6.0.2=pypi_0
qdldl=0.1.7.post5=pypi_0
readline=8.2=h5eee18b_0
rectpack=0.2.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.28.2=pypi_0
rich=13.4.2=pypi_0
ruff=0.9.10=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
scs=3.2.7.post2=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=60.2.0=pypi_0
shapely=2.1.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smplx=0.1.28=pypi_0
sniffio=1.3.1=pypi_0
sounddevice=0.5.1=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.46.1=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.2.18=pypi_0
timm=0.6.13=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
tomlkit=0.13.2=pypi_0
torch=2.3.1=pypi_0
torchaudio=2.3.1=pypi_0
torchmetrics=1.7.1=pypi_0
torchvision=0.18.1=pypi_0
tqdm=4.65.2=pypi_0
transformers=4.50.3=pypi_0
trimesh=4.6.6=pypi_0
triton=2.3.1=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
ultralytics=8.3.87=pypi_0
ultralytics-thop=2.0.14=pypi_0
urllib3=1.26.20=pypi_0
uvicorn=0.34.0=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.111=pypi_0
websockets=14.2=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xtcocotools=1.14.3=pypi_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.43.0=pypi_0
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
